Movie Recommender/recommender_models/requirements.txt
surprise 
python-dotenv 
azure-ai-ml 
mlflow 
azureml-mlflow